Remove deprecated model and controller concerns
