+++ lib/graphql_devise.rb
@@ -9,12 +9,7 @@ require 'zeitwerk'
 GraphQL::Field.accepts_definitions(authenticate: GraphQL::Define.assign_metadata_key(:authenticate))
 GraphQL::Schema::Field.accepts_definition(:authenticate)
 
-legacy_concerns = ['set_user_by_token', 'model']
 loader = Zeitwerk::Loader.for_gem
-
-legacy_concerns.each do |concern_name|
-  loader.ignore("#{__dir__}/graphql_devise/concerns/legacy/#{concern_name}.rb")
-end
 
 loader.collapse("#{__dir__}/graphql_devise/concerns")
 loader.collapse("#{__dir__}/graphql_devise/errors")
@@ -24,10 +19,6 @@ loader.inflector.inflect('error_codes' => 'ERROR_CODES')
 loader.inflector.inflect('supported_options' => 'SUPPORTED_OPTIONS')
 
 loader.setup
-
-legacy_concerns.each do |concern_name|
-  require_relative "graphql_devise/concerns/legacy/#{concern_name}"
-end
 
 ActionDispatch::Routing::Mapper.include(GraphqlDevise::RouteMounter)
