+++ spec/dummy/app/models/schema_user.rb
@@ -7,7 +7,7 @@ class SchemaUser < ApplicationRecord
          :validatable,
          :confirmable
 
-  include GraphqlDevise::Concerns::Model
+  include GraphqlDevise::Authenticatable
 
   validates :name, presence: true
 end
